Make the icons configurable
Icons can be configured now.
Two extra options are introduced:
icon, iconProvider

the icon can be any string that identifies the icon.
the iconProvider can be mdi (material design icon) or fa (font awesome)
The default iconprovider is mdi and default icon is mdp-event.
In case a wrong iconprovider is provided, the default icon will be used
which is the mdp-event icon from material design